Button sizes (#322)
~~This PR is not meant to be merged!~~
~~It's just a test to show what I meant in~~ #316

@biocom is mentioning
> Ideally it would respect the existing UNI override where it sets the inline style of 14px for the font-size. But that didn't seem to be the case and it seems our override method is the only way, given our general setup.

But I see it uses really this value, the thing is that is was 26px and not 14px. 
If I change it, all buttons become smaller:
![image](https://user-images.githubusercontent.com/2352112/112158057-196a4300-8be8-11eb-844a-36b66279a9a8.png)

Should we just add the size property to the buttons?
![image](https://user-images.githubusercontent.com/2352112/112158343-5b938480-8be8-11eb-87c6-9df1917b8e22.png)


Also, why in the original theme we didn't have this issue? the size of the button is way bigger in the cow theme than in the GP theme:
![image](https://user-images.githubusercontent.com/2352112/112158395-6b12cd80-8be8-11eb-9ccb-b15cffcdf100.png)

![Uploading image.png…]()


----- 


EDIT: Now i changed into a proposal that removes the sizes of the buttons, and instead adds  a new buttonSize property to create 3 different sizes for buttons: small, medium, big